Split rstest helper HIR visitor
Move HIR traversal, parameter lowering, and redacted fingerprint rendering
into `visitor.rs` so the driver stays within the repository's 400-line module
limit.

Keep the lint-pass lifecycle and call-site collection behaviour unchanged.
